5/14/2024_12:37
The login endpoints on the backend are all working. This iteration of the program is not using the MongoDB database only local storage. The functions are in place and need to be replaced with calls to the database. Moreover, the APIs are connected to EJS files and thus some of the controllers may need modifications to their responses to work better with React. The general framework is in place with minor modifications required.